Refactors bundle generation for deep/shallow modes
This commit introduces a refactoring of the bundle generation process to support both "deep" and "shallow" modes. It modifies the file lister to include basic file info, the file grouper to generate separate header files for deep and shallow bundles and adjusts the library template to incorporate includes based on the bundler mode. The bdfs module is updated to use base64 encoding for data transfer. This change allows for greater flexibility in how the bundle is created and used, with the goal of reducing memory consumption for shallow bundles. The VFS module is renamed to BDFS.