Implements shared buffer system for webview
Adds a shared buffer system to the webview to improve data transfer between the native code and the webview.

This system allows for efficient transfer of data, such as audio buffers or image data, by sharing memory rather than copying.

This change introduces a queue for managing shared buffers, a timer to process the queue, and mechanisms for the webview to request and receive these buffers. Also fixes a bug where plugin parameters weren't being sent